Built-in-Tools/remote-mcp-agent/python/requirements.txt
openai>=1.60.0
python-dotenv>=1.0.0
